move Teradata tests
