+++ tests/Functional/Teradata/Table/TeradataTableQueryBuilderTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Keboola\TableBackendUtils\Functional\Table\Teradata;
+namespace Tests\Keboola\TableBackendUtils\Functional\Teradata\Table;
 
 use Doctrine\DBAL\Exception as DBALException;
 use Keboola\TableBackendUtils\Column\ColumnCollection;
